intermediate step for refactoring
