+++ core/src/main/java/org/apache/sling/testing/mock/osgi/config/annotations/ApplyConfig.java
@@ -24,10 +24,11 @@ import java.lang.annotation.RetentionPolicy;
 
 /**
  * Defines an instance of an OSGi R7 Component Property Type as a combination of a {@link Class} and an array of strings
- * defining property values in the form expected by {@link Component#property()}. This provides both runtime retention
- * for OSGi config annotations that do not have {@link RetentionPolicy#RUNTIME}, allowing for simple construction
- * through reflection for explicit passing to SCR component constructors and lifecycle methods, as well as repeatability
- * to support defining sequenced, heterogeneous lists of desired types on any single {@link java.lang.reflect.AnnotatedElement}.
+ * defining property values in the form expected by {@link org.osgi.service.component.annotations.Component#property()}.
+ * This provides both runtime retention for OSGi config annotations that do not have {@link RetentionPolicy#RUNTIME},
+ * allowing for simple construction through reflection for explicit passing to SCR component constructors and lifecycle
+ * methods, as well as repeatability to support defining sequenced, heterogeneous lists of desired types on any single
+ * {@link java.lang.reflect.AnnotatedElement}.
  *
  * @see <a href="https://docs.osgi.org/specification/osgi.cmpn/7.0.0/service.component.html#service.component-component.property.types">Component Property Types</a>
  */
@@ -45,23 +46,23 @@ public @interface ApplyConfig {
     /**
      * Optionally specify a configuration pid to load, any defined properties of which will override annotation defaults
      * and values specified by {@link #property()}. In order to specify the name of the {@link #component()} class as a
-     * configuration PID, set this value to {@link Component#NAME}. The default value is the empty string, which
-     * skips loading any configuration from ConfigurationAdmin.
+     * configuration PID, set this value to {@link org.osgi.service.component.annotations.Component#NAME}. The default
+     * value is the empty string, which skips loading any configuration from ConfigurationAdmin.
      *
      * @return a configuration pid, or an empty string
      */
     String pid() default "";
 
     /**
-     * When {@link #pid()} is set to {@link Component#NAME}, set this attribute to a class whose name should be used
-     * instead.
+     * When {@link #pid()} is set to {@link org.osgi.service.component.annotations.Component#NAME}, set this attribute
+     * to a class whose name should be used instead.
      *
      * @return the configurable component class
      */
     Class<?> component() default Object.class;
 
     /**
-     * Treat like {@link Component#property()}.
+     * Treat like {@link org.osgi.service.component.annotations.Component#property()}.
      *
      * @return osgi component properties
      */
